refactor: move things + fix tests accordingly to prev changes
