+++ raspberrypi_central/smart-camera/app/camera/camera_factory.py
@@ -1,8 +1,8 @@
 from .camera import Camera
 import os
-from camera.detect_motion import DetectPeople
-from .camera_analyze import CameraAnalyzeObject
+from object_detection.detect_motion import DetectPeople
+from camera_analyze.camera_analyzer import CameraAnalyzer
 
 
-def camera_factory(get_mqtt_client, camera_analyze_object: CameraAnalyzeObject) -> Camera:
+def camera_factory(get_mqtt_client, camera_analyze_object: CameraAnalyzer) -> Camera:
     return Camera(camera_analyze_object, DetectPeople(), get_mqtt_client, os.environ['DEVICE_ID'])
