+++ raspberrypi_central/webapp/app/alarm/communication/alarm_consts.py
@@ -1,0 +1,7 @@
+from enum import Enum, unique
+
+
+@unique
+class ROITypes(Enum):
+    RECTANGLES = 'rectangles'
+    FULL = 'full'
